py-ai/pyproject.toml
[project]
name = "py-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.4.2",
]
